The HybridFPGrowth algorithm is roughly ready. Anyway it still not efficient to handle huge size of data because it seems not be able to reduce computation when matching records(/transactions) one by one to existing F2Itemsets and guessing corresponding FiItemsets.
I'm still thinking about some new way (for example, we can make FPGrowth
 really distributed by aggregating whenever a global step in FPGrowth is
done distributed on each sub-fptree on each node, that may cause too many
synchronization during the computation and may make the algorithm slow,
anyway that should make the algorithm really distributed and scalable. It
would be perfect if we can even optimize and speed up it) to make the
 algorithm really scalable and balanced so that to handle huge data size.
We need more time and careful thinking.